Catch SIGINT and SIGTERM (#11)
Recently we've hit the situation a couple of times where systemd-oomd
would kill solbuild during an intensive build which would completely
nuke all the packages the local repository. A lot of work and time can be lost
due to this so ensure we cleanup gracefully in such a situation.

Not sure whether this will completely solve the issue but it makes sense
to cleanup properly in such a situation as it is very difficult to
manually make solbuild usable again if it receives a SIGINT or SIGTERM
during a build without rebooting.