cli/delete_cache: Show disk usage of solbuild caches
`solbuild dc --sizes` will now show the disk usage of the various
solbuild caches as well as the total disk usage.

When deleting cache it will now additionally show how much disk space
was restored.

Computing the disk usage when deleting the cache adds quite a lot of
additional file stats, this may be removed in the future depending on
performance.