cli/build: Handle all option variables for tmpfs and memory size
Previously the config options for tmpfs and memory size where not
properly respected. For example, if you enabled tmpfs via the cli
but didn't specify a memory size it would not check for a memory size
in the config file. Now, all variations are accounted for.

Additionally, a new function `validMemSize` will check the memory
size used is valid before attempting to use it.