[WIP] Try to get go-git working how we want it
This tries to make go-git work with getting a shallow clone of a specific tag or commit. In its current state, getting a commit might work. Getting a tag, however, does not. Trying to resolve a tag ends up with a hash that doesn't exist when you try to fetch it. I don't know if I'm doing something wrong, go-git is doing something wrong, or a combination of both.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>