Add Accept header to HTTP requests
While debugging why netfilter.org was returning a 403 with the Solbuild client I noticed that wget (which succeeded) was sending the `Accept: */*` header. When I added it to Solbuild it started working. This header should be safe to send in all circumstances as it merely indicates that we're willing to accept any content-type (which we are).

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>