Add version to user-agent
The [HTTP User-Agent](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/User-Agent) specification requires that the user-agent has a version in it. Let's add this to hopefully improve success of this client with HTTP servers that require it.

Also, since we now reference the version from multiple places let's move configuration of it to the makefile and not have to duplicate references to it. Create a util package to avoid having to introduce a circular dependency (or to have it in the source package where it doesn't make sense)

Signed-off-by: Reilly Brogan <reilly@reillybrogan.com>