Use eopkg.bin to build (except for BuildXML)
In various tests, this is showing a speedup for install operations on the
order of 20-25% (very noticeable for e.g. the ISO generation scripts).

This is also part of the preparation for dropping the py2 pisi package
(which owns the `eopkg` symlink at the time of this writing).

Note that this will keep using `eopkg.py2` for building old pspec.xml
packages for now; switching to using the explicit version avoids
solbuild having to rely on which package owns the `eopkg` symlink.

When the epoch bump happens and eopkg.py2 is dropped from the repo,
BuildXML needs to be changed to use `eopkg.py3`.

Note that ypkg already rundeps on python-eopkg, which contains eopkg.py3.

Signed-off-by: Rune Morling <ermo@serpentos.com>