builder: Detect which ypkg command to use
Once https://github.com/getsolus/ypkg/pull/114 lands, the CLI will be
changed. Instead of having separate scripts that call into `ypkg2`,
there will be a single entry-point with subcommands.

This commit checks for the presence of the old scripts, and uses the new
command syntax if they are not available.

Signed-off-by: Evan Maddock <maddock.evan@vivaldi.net>